Improve shortcuts response
