+++ versions/v4/interfaces/models/line_estimations_interface.ts
@@ -4,3 +4,10 @@ export type LineEstimations = [
   minutes1: number | null,
   minutes2: number | null,
 ];
+
+export interface ShortcutsLineEstimations {
+  label: string;
+  destination: string;
+  minutes1: number | null;
+  minutes2: number | null;
+}
