+++ versions/v4/interfaces/responses/stop_estimations_interface.ts
@@ -4,14 +4,3 @@ export type StopEstimations = [
   lineEstimations: LineEstimations[],
   lineLabelsOrStopNames?: string[],
 ];
-
-export interface ShortcutsStopEstimations {
-  [key: string]: ShortcutsLineEstimations;
-}
-
-export interface ShortcutsLineEstimations {
-  label: string;
-  destination: string;
-  minutes1: number | null;
-  minutes2: number | null;
-}
